feat: refactor IngestItemsList and place StepActionButton component in separate file
